pyproject.toml
[project]
name = "spectre-terminal-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "pytest>=9.0.0",
    "pytest-html>=4.1.1",
    "python-dotenv>=1.2.1",
    "pywinauto>=0.6.9",
]
